+++ src/PurgeAssetUrlJob.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Noo\CraftBunnyPurge;
+namespace Noo\CraftCdnAssetPurge;
 
 use craft\queue\BaseJob;
 
@@ -10,11 +10,11 @@ class PurgeAssetUrlJob extends BaseJob
 
     public function execute($queue): void
     {
-        BunnyPurge::getInstance()->purgeService->purgeUrls([$this->url]);
+        CdnAssetPurge::getInstance()->purgeService->purgeUrls([$this->url]);
     }
 
     protected function defaultDescription(): ?string
     {
-        return "Purging Bunny CDN cache for {$this->url}";
+        return "Purging CDN cache for {$this->url}";
     }
 }